message

+++ packageKPS.py
@@ -99,12 +99,12 @@ def get_data_from_json(kdata):
     matching_columns = kdata.columns[kdata.columns.str.startswith("report.kad.81.")]
     print(matching_columns)
 
-    kdata[matching_columns] = kdata[matching_columns].astype(int)
+    kdata[matching_columns] = kdata[matching_columns].astype(float)
 
     kpdf['D26'] = kdata.apply(lambda row: calculate_d26_d27(row, matching_columns), axis=1)
     #search for kad starts from .56
     matching_columns2 = kdata.columns[kdata.columns.str.startswith("report.kad.56.")]
-    kdata[matching_columns2] = kdata[matching_columns2].astype(int)
+    kdata[matching_columns2] = kdata[matching_columns2].astype(float)
     
     kpdf['D27'] = kdata.apply(lambda row: calculate_d26_d27(row, matching_columns2), axis=1)
